Restructured JDFTXOutfile to be an inherited list of JDFTXOutfileSlice's (which is the same as JDFTXOutfile was but only initializes from an out file slice). All previous functionality is retained by calling and setting attributes to the most recent JDFTXOutfileSlice
